Refactor `parse_int64_bytes` to be part of utils. #7514
In `submit_utils.cpp`, we had the useful `parse_int64_bytes` function
that took a config value and converted it to a `uint64_t` value
(correctly handling suffixes like MB, GB).  This moves the function
from `submit_utils` to `metric_utils` so others can benefit as well.
